+++ CloudShell.ResourceModel/ResourceCapabilityAttributes.cs
@@ -1,0 +1,10 @@
+namespace CloudShell.ResourceModel;
+
+public interface IResourceCapabilityAttributeProvider
+{
+    ResourceCapabilityId CapabilityId { get; }
+
+    IReadOnlyDictionary<ResourceAttributeId, ResourceAttributeDefinition> AttributeDefinitions { get; }
+
+    IReadOnlyDictionary<ResourceAttributeValueShapeId, ResourceAttributeValueShapeDefinition> AttributeValueShapes { get; }
+}